Add product-hashcode & calculate heap

+++ models/product.go
@@ -4,13 +4,13 @@ type Product struct {
 	Id            int     `json:"id" gorm:"autoIncrement;column:id"`
 	SourceId      int     `json:"source_id" gorm:"column:sourceid"`
 	ProductTitle  string  `json:"product_title" gorm:"column:title"`
-	ProductPrice  float64 `json:"product_price" gorm:"column:price"`
-	StockQuantity int     `json:"stock_quantity" gorm:"column:stock"`
+	ProductPrice  float32 `json:"product_price" gorm:"column:price"`
+	StockQuantity uint16  `json:"stock_quantity" gorm:"column:stock"`
 }
 
 type CreateProduct struct {
 	SourceId      int     `json:"id" gorm:"column:sourceid"`
 	ProductTitle  string  `json:"product_title" gorm:"column:title"`
-	ProductPrice  float64 `json:"product_price" gorm:"column:price"`
-	StockQuantity int     `json:"stock_quantity" gorm:"column:stock"`
+	ProductPrice  float32 `json:"product_price" gorm:"column:price"`
+	StockQuantity uint16  `json:"stock_quantity" gorm:"column:stock"`
 }
